api_backend/requirements.txt
flask 
mysql-connector-python

